Make breaking news example use a sigil
